Limits to data shown
Data will only be shown from the start date of the billing file. In most cases that means the first day the CUR is generated, but for Eco Reports it may mean first day the modifed CUR is generated. If you need more, retroactive data, please submit a support ticket.